multiDeleting

+++ src/ServletDelete.java
@@ -2,6 +2,8 @@ import com.*;
 
 import java.io.IOException;
 import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.Arrays;
 
 /**
  * Created by fil-n on 22.04.17.
@@ -20,8 +22,12 @@ public class ServletDelete extends javax.servlet.http.HttpServlet {
         String table = request.getParameter("table");
         try {
             if (table.equals("film")){
-                int id = Integer.valueOf(request.getParameter("id"));
-                FilmDAO.deleteFilm(id);
+                    if (!request.getParameter("id").equals("")){
+                    int[] arDel = Arrays.stream(request.getParameter("id").split(",")).mapToInt(Integer::parseInt).toArray();
+                    for (int id: arDel ) {
+                        FilmDAO.deleteFilm(id);
+                    }
+                }
             }
             if (table.equals("frame")){
                 int id = Integer.valueOf(request.getParameter("id"));
